requirements.txt
# API
fastapi==0.115.6
uvicorn[standard]==0.32.1

# RAG & Embeddings
openai==1.58.1
pinecone==5.4.2
langchain-text-splitters==0.3.3
pypdf==5.1.0

# Data & Config
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
pandas==2.2.3

# Testing
pytest==8.3.4
httpx==0.28.1
